requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10